python3.12/django/app_09/pyproject.toml
[tool.poetry]
name = "app-09"
version = "0.1.0"
description = ""
authors = ["noskill <golanghack@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
django-rest-framework = "^0.1.0"
celery = "^5.4.0"
redis = "^5.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
